+++ include/highfive/bits/H5Node_traits_misc.hpp
@@ -36,12 +36,14 @@ inline DataSet
 NodeTraits<Derivate>::createDataSet(const std::string& dataset_name,
                                     const DataSpace& space,
                                     const DataType& dtype,
-                                    const DataSetCreateProps& createProps) {
+                                    const DataSetCreateProps& createProps,
+                                    const DataSetAccessProps& accessProps)
+{
     DataSet set;
     if ((set._hid = H5Dcreate2(static_cast<Derivate*>(this)->getId(),
                                dataset_name.c_str(), dtype._hid, space._hid,
                                H5P_DEFAULT, createProps.getId(),
-                               H5P_DEFAULT)) < 0) {
+                               accessProps.getId())) < 0) {
         HDF5ErrMapper::ToException<DataSetException>(
             std::string("Unable to create the dataset \"") + dataset_name +
             "\":");
@@ -54,17 +56,20 @@ template <typename Type>
 inline DataSet
 NodeTraits<Derivate>::createDataSet(const std::string& dataset_name,
                                     const DataSpace& space,
-                                    const DataSetCreateProps& createProps) {
+                                    const DataSetCreateProps& createProps,
+                                    const DataSetAccessProps& accessProps)
+{
     return createDataSet(dataset_name, space, AtomicType<Type>(),
-                         createProps);
+                         createProps, accessProps);
 }
 
 template <typename Derivate>
 inline DataSet
-NodeTraits<Derivate>::getDataSet(const std::string& dataset_name) const {
+NodeTraits<Derivate>::getDataSet(const std::string& dataset_name,
+                                 const DataSetAccessProps& accessProps) const {
     DataSet set;
     if ((set._hid = H5Dopen2(static_cast<const Derivate*>(this)->getId(),
-                             dataset_name.c_str(), H5P_DEFAULT)) < 0) {
+                             dataset_name.c_str(), accessProps.getId())) < 0) {
         HDF5ErrMapper::ToException<DataSetException>(
             std::string("Unable to open the dataset \"") + dataset_name +
             "\":");